Перенос моделей настроек и таймера в сервисы. Удаление стартового окна

+++ FocusCycle/Models/TimerSettings.cs
@@ -12,6 +12,6 @@ namespace FocusCycle.Models
         public TimeSpan BreakTime { get; set; } = TimeSpan.FromSeconds(15);
 
         [Key(2)]
-        public double Volume { get; set; } = 1.0;
+        public float Volume { get; set; } = 1.0F;
     }
 }
